Reworks for geofilter serialization work
- Remove serde module. Put code into diff_count.rs.
- Remove bonus comments
- Remove panic-y big endian serialization stuff.
- Squish together lines
